Trait order_book::pallet::Config

pub trait Config: Config + Config + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type OrderId: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + AtLeast32BitUnsigned + Copy + Ord + PartialEq + Eq + MaxEncodedLen + TypeInfo;
    type MaxOpenedLimitOrdersForAllOrderBooksPerUser: Get<u32>;
    type MaxLimitOrdersForPrice: Get<u32>;
    type MaxSidePrices: Get<u32>;
    type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, Self::AssetId, DispatchError>;
    type AssetInfoProvider: AssetInfoProvider<Self::AssetId, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>;
    type DexInfoProvider: DexInfoProvider<Self::DEXId, DEXInfo<Self::AssetId>>;
    type WeightInfo: WeightInfo;

    const MAX_ORDER_LIFETIME: Self::Moment;
    const MAX_OPENED_LIMIT_ORDERS_COUNT: u32;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type OrderId: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + AtLeast32BitUnsigned + Copy + Ord + PartialEq + Eq + MaxEncodedLen + TypeInfo

type MaxOpenedLimitOrdersForAllOrderBooksPerUser: Get<u32>

type MaxLimitOrdersForPrice: Get<u32>

type MaxSidePrices: Get<u32>

type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, Self::AssetId, DispatchError>

type AssetInfoProvider: AssetInfoProvider<Self::AssetId, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>

type DexInfoProvider: DexInfoProvider<Self::DEXId, DEXInfo<Self::AssetId>>

type WeightInfo: WeightInfo

Required Associated Constants

const MAX_ORDER_LIFETIME: Self::Moment

const MAX_OPENED_LIMIT_ORDERS_COUNT: u32