Trait order_book::traits::DataLayer

pub trait DataLayer<T>where
    T: Config,{

    // Required methods
    fn get_limit_order(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        order_id: T::OrderId
    ) -> Result<LimitOrder<T>, DispatchError>;
    fn get_all_limit_orders(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> Vec<LimitOrder<T>> ;
    fn insert_limit_order(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        order: LimitOrder<T>
    ) -> Result<(), DispatchError>;
    fn update_limit_order_amount(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        order_id: T::OrderId,
        new_amount: OrderVolume
    ) -> Result<(), DispatchError>;
    fn delete_limit_order(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        order_id: T::OrderId
    ) -> Result<(), DispatchError>;
    fn get_bids(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        price: &OrderPrice
    ) -> Option<PriceOrders<T::OrderId, T::MaxLimitOrdersForPrice>>;
    fn is_bid_price_full(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        price: &OrderPrice
    ) -> Option<bool>;
    fn get_asks(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        price: &OrderPrice
    ) -> Option<PriceOrders<T::OrderId, T::MaxLimitOrdersForPrice>>;
    fn is_ask_price_full(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        price: &OrderPrice
    ) -> Option<bool>;
    fn get_aggregated_bids(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> MarketSide<T::MaxSidePriceCount>;
    fn get_aggregated_bids_len(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> Option<usize>;
    fn best_bid(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> Option<(OrderPrice, OrderVolume)>;
    fn get_aggregated_asks(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> MarketSide<T::MaxSidePriceCount>;
    fn get_aggregated_asks_len(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> Option<usize>;
    fn best_ask(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> Option<(OrderPrice, OrderVolume)>;
    fn get_user_limit_orders(
        &mut self,
        account: &T::AccountId,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> Option<UserOrders<T::OrderId, T::MaxOpenedLimitOrdersPerUser>>;
    fn get_all_user_limit_orders(
        &mut self,
        account: &T::AccountId
    ) -> BTreeMap<OrderBookId<AssetIdOf<T>, T::DEXId>, UserOrders<T::OrderId, T::MaxOpenedLimitOrdersPerUser>>;
    fn is_user_limit_orders_full(
        &mut self,
        account: &T::AccountId,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>
    ) -> Option<bool>;

    // Provided method
    fn get_limit_orders_by_price(
        &mut self,
        order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>,
        side: PriceVariant,
        price: &OrderPrice
    ) -> Option<PriceOrders<T::OrderId, T::MaxLimitOrdersForPrice>> { ... }

}

This trait is used by Order Book as a storage to get limit orders and their derived data and to change them

Required Methods

fn get_limit_order( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, order_id: T::OrderId ) -> Result<LimitOrder<T>, DispatchError>

Returns the limit order if exists, otherwise returns error.

fn get_all_limit_orders( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> Vec<LimitOrder<T>>

Returns all limit orders of order book

fn insert_limit_order( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, order: LimitOrder<T> ) -> Result<(), DispatchError>

Inserts limit order consistently in all necessary storages. Must check before call. If returns error, it means we have problems with data consistency. If order_id already exists - returns error. Use update_limit_order to update the existing order.

fn update_limit_order_amount( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, order_id: T::OrderId, new_amount: OrderVolume ) -> Result<(), DispatchError>

Updates the amount of the limit order. If order doesn’t exist - return error

fn delete_limit_order( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, order_id: T::OrderId ) -> Result<(), DispatchError>

Deletes limit order consistently from all necessary storages. If order doesn’t exist - return error Must check before call. If returns error, it means we have problems with data consistency.

fn get_bids( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, price: &OrderPrice ) -> Option<PriceOrders<T::OrderId, T::MaxLimitOrdersForPrice>>

Returns order ids of orders inside the bid price

fn is_bid_price_full( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, price: &OrderPrice ) -> Option<bool>

Returns whether there is no place for orders inside the bid price. None if no entry for the price exists

fn get_asks( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, price: &OrderPrice ) -> Option<PriceOrders<T::OrderId, T::MaxLimitOrdersForPrice>>

Returns order ids of orders inside the ask price

fn is_ask_price_full( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, price: &OrderPrice ) -> Option<bool>

Returns whether there is no place for orders inside the ask price. None if no entry for the price exists

fn get_aggregated_bids( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> MarketSide<T::MaxSidePriceCount>

Returns all bid prices with their volumes

fn get_aggregated_bids_len( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> Option<usize>

Length of aggregated asks list for the order book. None if the value is not present in the storage. Ignores default value, so unwrap_or(0) usually makes sense.

fn best_bid( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> Option<(OrderPrice, OrderVolume)>

fn get_aggregated_asks( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> MarketSide<T::MaxSidePriceCount>

Returns all ask prices with their volumes

fn get_aggregated_asks_len( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> Option<usize>

Length of aggregated asks list for the order book. None if the value is not present in the storage. Ignores default value, so unwrap_or(0) usually makes sense.

fn best_ask( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> Option<(OrderPrice, OrderVolume)>

fn get_user_limit_orders( &mut self, account: &T::AccountId, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> Option<UserOrders<T::OrderId, T::MaxOpenedLimitOrdersPerUser>>

Returns order ids of user from the order book with order_book_id

fn get_all_user_limit_orders( &mut self, account: &T::AccountId ) -> BTreeMap<OrderBookId<AssetIdOf<T>, T::DEXId>, UserOrders<T::OrderId, T::MaxOpenedLimitOrdersPerUser>>

Returns order ids of user from all order books

fn is_user_limit_orders_full( &mut self, account: &T::AccountId, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId> ) -> Option<bool>

Returns whether there is no place for the user’s orders in the order book. None if there is no entry tracking orders of the user in the order book.

Provided Methods

fn get_limit_orders_by_price( &mut self, order_book_id: &OrderBookId<AssetIdOf<T>, T::DEXId>, side: PriceVariant, price: &OrderPrice ) -> Option<PriceOrders<T::OrderId, T::MaxLimitOrdersForPrice>>

Returns order ids of orders inside the bid or ask price