Trait order_book::pallet::Config

pub trait Config: Config + Config + Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type OrderId: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + AtLeast32BitUnsigned + Copy + Ord + PartialEq + Eq + MaxEncodedLen + TypeInfo;
    type Locker: CurrencyLocker<Self::AccountId, Self::AssetId, Self::DEXId, DispatchError>;
    type Unlocker: CurrencyUnlocker<Self::AccountId, Self::AssetId, Self::DEXId, DispatchError>;
    type Scheduler: ExpirationScheduler<Self::BlockNumber, OrderBookId<Self::AssetId, Self::DEXId>, Self::DEXId, Self::OrderId, DispatchError>;
    type Delegate: Delegate<Self::AccountId, Self::AssetId, Self::OrderId, Self::DEXId>;
    type MaxOpenedLimitOrdersPerUser: Get<u32>;
    type MaxLimitOrdersForPrice: Get<u32>;
    type MaxSidePriceCount: Get<u32>;
    type MaxExpiringOrdersPerBlock: Get<u32>;
    type MaxExpirationWeightPerBlock: Get<Weight>;
    type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, Self::AssetId, DispatchError>;
    type TradingPairSourceManager: TradingPairSourceManager<Self::DEXId, Self::AssetId>;
    type AssetInfoProvider: AssetInfoProvider<Self::AssetId, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>;
    type SyntheticInfoProvider: SyntheticInfoProvider<Self::AssetId>;
    type DexInfoProvider: DexInfoProvider<Self::DEXId, DEXInfo<Self::AssetId>>;
    type Time: Time;
    type ParameterUpdateOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Either<Self::AccountId, ()>>;
    type StatusUpdateOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = ()>;
    type RemovalOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = ()>;
    type WeightInfo: WeightInfo;

    const MAX_ORDER_LIFESPAN: MomentOf<Self>;
    const MIN_ORDER_LIFESPAN: MomentOf<Self>;
    const MILLISECS_PER_BLOCK: MomentOf<Self>;
    const MAX_PRICE_SHIFT: Perbill;
    const SOFT_MIN_MAX_RATIO: usize;
    const HARD_MIN_MAX_RATIO: usize;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type OrderId: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + AtLeast32BitUnsigned + Copy + Ord + PartialEq + Eq + MaxEncodedLen + TypeInfo

type Locker: CurrencyLocker<Self::AccountId, Self::AssetId, Self::DEXId, DispatchError>

type Unlocker: CurrencyUnlocker<Self::AccountId, Self::AssetId, Self::DEXId, DispatchError>

type Scheduler: ExpirationScheduler<Self::BlockNumber, OrderBookId<Self::AssetId, Self::DEXId>, Self::DEXId, Self::OrderId, DispatchError>

type Delegate: Delegate<Self::AccountId, Self::AssetId, Self::OrderId, Self::DEXId>

type MaxOpenedLimitOrdersPerUser: Get<u32>

type MaxLimitOrdersForPrice: Get<u32>

type MaxSidePriceCount: Get<u32>

type MaxExpiringOrdersPerBlock: Get<u32>

type MaxExpirationWeightPerBlock: Get<Weight>

type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, Self::AssetId, DispatchError>

type TradingPairSourceManager: TradingPairSourceManager<Self::DEXId, Self::AssetId>

type AssetInfoProvider: AssetInfoProvider<Self::AssetId, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>

type SyntheticInfoProvider: SyntheticInfoProvider<Self::AssetId>

type DexInfoProvider: DexInfoProvider<Self::DEXId, DEXInfo<Self::AssetId>>

type Time: Time

type ParameterUpdateOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Either<Self::AccountId, ()>>

type StatusUpdateOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = ()>

type RemovalOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = ()>

type WeightInfo: WeightInfo

Required Associated Constants

const MAX_ORDER_LIFESPAN: MomentOf<Self>

const MIN_ORDER_LIFESPAN: MomentOf<Self>

const MILLISECS_PER_BLOCK: MomentOf<Self>

const MAX_PRICE_SHIFT: Perbill

const SOFT_MIN_MAX_RATIO: usize

The soft ratio between min & max order amounts. In particular, it defines the optimal number of limit orders that could be executed by one big market order in one block.

const HARD_MIN_MAX_RATIO: usize

The soft ratio between min & max order amounts. In particular, it defines the max number of limit orders that could be executed by one big market order in one block.